pyproject.toml
[tool.poetry]
name = "tensorflow-test"
version = "0.1.0"
description = ""
authors = ["Fhilipe Coelho <fhilipecoelho.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
matplotlib = "^3.9.2"
scikit-learn = "^1.5.2"
joblib = "^1.4.2"
pydicom = "^3.0.1"
scikit-image = "^0.24.0"
opencv-python = "^4.10.0.84"
opencv-cuda = "^0.0.2"
cupy-cuda11x = "^13.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
